Allow numeric subaccount
When a numeric subaccount is passed in the params list, it is being
interpreted as int because the 'exec' method needs to parse the
parameter list.
This is kind of a workaround since the parameter will still be
converted to int and then formatted back as string when inserting
in the graphql parameterer.

Without this workaround, the grapqhl API will return an error
informing that the 'username' cannot be an integer.